+++ packages/yew-macro/tests/html_macro/block-fail.rs
