+++ packages/yew-macro/tests/html_macro/block-pass.rs
