+++ packages/yew-macro/tests/html_macro/component-pass.rs
@@ -1,60 +1,8 @@
 #![recursion_limit = "256"]
 
-use std::marker::PhantomData;
 use yew::html::ChildrenRenderer;
 use yew::prelude::*;
 use yew::virtual_dom::{VChild, VNode};
-
-pub struct Generic<T> {
-    marker: PhantomData<T>,
-}
-
-impl<T> Component for Generic<T>
-where
-    T: 'static,
-{
-    type Message = ();
-    type Properties = ();
-
-    fn create(_: Self::Properties, _: ComponentLink<Self>) -> Self {
-        unimplemented!()
-    }
-    fn update(&mut self, _: Self::Message) -> ShouldRender {
-        unimplemented!()
-    }
-    fn change(&mut self, _: Self::Properties) -> ShouldRender {
-        unimplemented!()
-    }
-    fn view(&self) -> Html {
-        unimplemented!()
-    }
-}
-
-pub struct Generic2<T1, T2> {
-    marker: PhantomData<(T1, T2)>,
-}
-
-impl<T1, T2> Component for Generic2<T1, T2>
-where
-    T1: 'static,
-    T2: 'static,
-{
-    type Message = ();
-    type Properties = ();
-
-    fn create(_: Self::Properties, _: ComponentLink<Self>) -> Self {
-        unimplemented!()
-    }
-    fn update(&mut self, _: Self::Message) -> ShouldRender {
-        unimplemented!()
-    }
-    fn change(&mut self, _: Self::Properties) -> ShouldRender {
-        unimplemented!()
-    }
-    fn view(&self) -> Html {
-        unimplemented!()
-    }
-}
 
 #[derive(Clone, Properties, Default)]
 pub struct ContainerProperties {
@@ -346,28 +294,6 @@ fn compile_pass() {
                         .collect::<VNode>()
                 }
             </div>
-        </>
-    };
-
-    html! {
-        <>
-            <Generic<String> />
-            <Generic<String> ></Generic<String>>
-            // nested generics
-            <Generic<Vec<String>> />
-            <Generic<Vec<String>>></ Generic<Vec<String>>>
-            // with lowercase type param
-            <Generic<usize> />
-            <Generic<usize>></Generic<usize>>
-            // with trailing comma
-            <Generic<String, > />
-            <Generic<String, >></Generic<String,>>
-            // with multiple type params
-            <Generic2<String, String> />
-            <Generic2<String, String>></Generic2<String, String>>
-            // with multiple type params and trailing comma
-            <Generic2<String, String, > />
-            <Generic2<String, String, >></Generic2<String, String, >>
         </>
     };
