+++ packages/yew-macro/tests/html_macro/component-unimplemented-fail.stderr
@@ -1,5 +1,5 @@
 error[E0599]: no function or associated item named `new` found for struct `yew::virtual_dom::vcomp::VChild<Unimplemented>` in the current scope
- --> $DIR/html-component-unimplemented-fail.rs:8:14
+ --> $DIR/component-unimplemented-fail.rs:8:14
   |
 5 | struct Unimplemented;
   | --------------------- doesn't satisfy `Unimplemented: yew::html::component::Component`
@@ -11,7 +11,7 @@ error[E0599]: no function or associated item named `new` found for struct `yew::
           `Unimplemented: yew::html::component::Component`
 
 error[E0277]: the trait bound `Unimplemented: yew::html::component::Component` is not satisfied
- --> $DIR/html-component-unimplemented-fail.rs:8:14
+ --> $DIR/component-unimplemented-fail.rs:8:14
   |
 8 |     html! { <Unimplemented /> };
   |              ^^^^^^^^^^^^^ the trait `yew::html::component::Component` is not implemented for `Unimplemented`
