+++ packages/yew-macro/tests/html_macro/generic-component-fail.stderr
@@ -1,0 +1,17 @@
+error: this opening tag has no corresponding closing tag
+  --> $DIR/generic-component-fail.rs:58:13
+   |
+58 |     html! { <Generic<String>> };
+   |             ^^^^^^^^^^^^^^^^^
+
+error: this closing tag has no corresponding opening tag
+  --> $DIR/generic-component-fail.rs:59:30
+   |
+59 |     html! { <Generic<String>></Generic> };
+   |                              ^^^^^^^^^^
+
+error: this closing tag has no corresponding opening tag
+  --> $DIR/generic-component-fail.rs:60:30
+   |
+60 |     html! { <Generic<String>></Generic<Vec<String>>> };
+   |                              ^^^^^^^^^^^^^^^^^^^^^^^
