+++ packages/yew-macro/tests/html_macro/iterable-fail.rs
