+++ packages/yew-macro/tests/html_macro/iterable-fail.stderr
@@ -1,11 +1,11 @@
 error: expected an expression after the keyword `for`
- --> $DIR/html-iterable-fail.rs:4:13
+ --> $DIR/iterable-fail.rs:4:13
   |
 4 |     html! { for };
   |             ^^^
 
 error[E0277]: `()` is not an iterator
- --> $DIR/html-iterable-fail.rs:5:17
+ --> $DIR/iterable-fail.rs:5:17
   |
 5 |     html! { for () };
   |                 ^^ `()` is not an iterator
@@ -14,7 +14,7 @@ error[E0277]: `()` is not an iterator
   = note: required by `std::iter::IntoIterator::into_iter`
 
 error[E0277]: `()` is not an iterator
- --> $DIR/html-iterable-fail.rs:6:17
+ --> $DIR/iterable-fail.rs:6:17
   |
 6 |     html! { for {()} };
   |                 ^^^^ `()` is not an iterator
@@ -23,7 +23,7 @@ error[E0277]: `()` is not an iterator
   = note: required by `std::iter::IntoIterator::into_iter`
 
 error[E0277]: `()` doesn't implement `std::fmt::Display`
- --> $DIR/html-iterable-fail.rs:7:17
+ --> $DIR/iterable-fail.rs:7:17
   |
 7 |     html! { for Vec::<()>::new().into_iter() };
   |                 ^^^ `()` cannot be formatted with the default formatter
@@ -37,7 +37,7 @@ error[E0277]: `()` doesn't implement `std::fmt::Display`
   = note: required by `std::iter::Iterator::collect`
 
 error[E0277]: `()` doesn't implement `std::fmt::Display`
-  --> $DIR/html-iterable-fail.rs:10:17
+  --> $DIR/iterable-fail.rs:10:17
    |
 10 |     html! { for empty };
    |                 ^^^^^ `()` cannot be formatted with the default formatter
@@ -51,7 +51,7 @@ error[E0277]: `()` doesn't implement `std::fmt::Display`
    = note: required by `std::iter::Iterator::collect`
 
 error[E0277]: `()` doesn't implement `std::fmt::Display`
-  --> $DIR/html-iterable-fail.rs:13:17
+  --> $DIR/iterable-fail.rs:13:17
    |
 13 |     html! { for empty.iter() };
    |                 ^^^^^ `()` cannot be formatted with the default formatter
@@ -66,14 +66,14 @@ error[E0277]: `()` doesn't implement `std::fmt::Display`
    = note: required by `std::iter::Iterator::collect`
 
 error[E0277]: `()` is not an iterator
-  --> $DIR/html-iterable-fail.rs:18:19
+  --> $DIR/iterable-fail.rs:18:19
    |
 18 |             { for () }
    |                   ^^ `()` is not an iterator
    |
-  ::: $WORKSPACE/packages/yew/src/utils.rs:50:9
+  ::: $WORKSPACE/packages/yew/src/utils.rs
    |
-50 |     IT: IntoIterator<Item = T>,
+   |     IT: IntoIterator<Item = T>,
    |         ---------------------- required by this bound in `yew::utils::into_node_iter`
    |
    = help: the trait `std::iter::Iterator` is not implemented for `()`
