+++ packages/yew-macro/tests/html_macro/list-fail.stderr
@@ -1,71 +1,71 @@
 error: this opening fragment has no corresponding closing fragment
- --> $DIR/html-list-fail.rs:4:13
+ --> $DIR/list-fail.rs:4:13
   |
 4 |     html! { <> };
   |             ^^
 
 error: this closing fragment has no corresponding opening fragment
- --> $DIR/html-list-fail.rs:5:13
+ --> $DIR/list-fail.rs:5:13
   |
 5 |     html! { </> };
   |             ^^^
 
 error: this opening fragment has no corresponding closing fragment
- --> $DIR/html-list-fail.rs:6:15
+ --> $DIR/list-fail.rs:6:15
   |
 6 |     html! { <><> };
   |               ^^
 
 error: this closing fragment has no corresponding opening fragment
- --> $DIR/html-list-fail.rs:7:13
+ --> $DIR/list-fail.rs:7:13
   |
 7 |     html! { </></> };
   |             ^^^
 
 error: this opening fragment has no corresponding closing fragment
- --> $DIR/html-list-fail.rs:8:13
+ --> $DIR/list-fail.rs:8:13
   |
 8 |     html! { <><></> };
   |             ^^
 
 error: only one root html element is allowed (hint: you can wrap multiple html elements in a fragment `<></>`)
- --> $DIR/html-list-fail.rs:9:18
+ --> $DIR/list-fail.rs:9:18
   |
 9 |     html! { <></><></> };
   |                  ^^^^^
 
 error: expected a valid html element
-  --> $DIR/html-list-fail.rs:10:15
+  --> $DIR/list-fail.rs:10:15
    |
 10 |     html! { <>invalid</> };
    |               ^^^^^^^
 
 error: expected an expression following this equals sign
-  --> $DIR/html-list-fail.rs:11:17
+  --> $DIR/list-fail.rs:11:17
    |
 11 |     html! { <key=></> };
    |                 ^^
 
 error: this closing tag has no corresponding opening tag
-  --> $DIR/html-list-fail.rs:12:36
+  --> $DIR/list-fail.rs:12:36
    |
 12 |     html! { <key="key".to_string()></key> };
    |                                    ^^^^^^
 
 error: only a single `key` prop is allowed on a fragment
-  --> $DIR/html-list-fail.rs:14:30
+  --> $DIR/list-fail.rs:14:30
    |
 14 |     html! { <key="first key" key="second key" /> };
    |                              ^^^
 
 error: fragments only accept the `key` prop
-  --> $DIR/html-list-fail.rs:15:14
+  --> $DIR/list-fail.rs:15:14
    |
 15 |     html! { <some_attr="test"></> };
    |              ^^^^^^^^^
 
 error: `key` does not support being used as an optional attribute
-  --> $DIR/html-list-fail.rs:17:14
+  --> $DIR/list-fail.rs:17:14
    |
 17 |     html! { <key?=None></> };
    |              ^^^^^
