+++ packages/yew-macro/tests/html_macro/node-fail.stderr
@@ -1,47 +1,47 @@
 error: only one root html element is allowed (hint: you can wrap multiple html elements in a fragment `<></>`)
- --> $DIR/html-node-fail.rs:4:21
+ --> $DIR/node-fail.rs:4:21
   |
 4 |     html! { "valid" "invalid" };
   |                     ^^^^^^^^^
 
 error: unexpected token
- --> $DIR/html-node-fail.rs:5:29
+ --> $DIR/node-fail.rs:5:29
   |
 5 |     html! { <span>{ "valid" "invalid" }</span> };
   |                             ^^^^^^^^^
 
 error: unsupported type
-  --> $DIR/html-node-fail.rs:10:14
+  --> $DIR/node-fail.rs:10:14
    |
 10 |     html! {  b'a' };
    |              ^^^^
 
 error: unsupported type
-  --> $DIR/html-node-fail.rs:11:14
+  --> $DIR/node-fail.rs:11:14
    |
 11 |     html! {  b"str" };
    |              ^^^^^^
 
 error: unsupported type
-  --> $DIR/html-node-fail.rs:12:22
+  --> $DIR/node-fail.rs:12:22
    |
 12 |     html! {  <span>{ b'a' }</span> };
    |                      ^^^^
 
 error: unsupported type
-  --> $DIR/html-node-fail.rs:13:22
+  --> $DIR/node-fail.rs:13:22
    |
 13 |     html! {  <span>{ b"str" }</span> };
    |                      ^^^^^^
 
 error[E0425]: cannot find value `invalid` in this scope
- --> $DIR/html-node-fail.rs:7:13
+ --> $DIR/node-fail.rs:7:13
   |
 7 |     html! { invalid };
   |             ^^^^^^^ not found in this scope
 
 error[E0277]: `()` doesn't implement `std::fmt::Display`
- --> $DIR/html-node-fail.rs:6:13
+ --> $DIR/node-fail.rs:6:13
   |
 6 |     html! { () };
   |             ^^ `()` cannot be formatted with the default formatter
@@ -53,7 +53,7 @@ error[E0277]: `()` doesn't implement `std::fmt::Display`
   = note: required by `std::convert::From::from`
 
 error[E0277]: `()` doesn't implement `std::fmt::Display`
-  --> $DIR/html-node-fail.rs:17:9
+  --> $DIR/node-fail.rs:17:9
    |
 17 |         not_node()
    |         ^^^^^^^^^^ `()` cannot be formatted with the default formatter
